changed column definition of some fields so that they work with hsql, which we use in maven-embedded-tomcat way of running WISE.\nConverted AnnouncementImpl.java from dos to unix format.

+++ src/main/java/org/wise/portal/domain/announcement/impl/AnnouncementImpl.java
@@ -68,7 +68,7 @@ public class AnnouncementImpl implements Announcement, Comparable<Announcement>{
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.AUTO)
-	@Column(name = "id", columnDefinition = "mediumint")
+	@Column(name = "id", columnDefinition = "int")
 	private Integer id;
 
 	/**
